requirements.txt
fastapi==0.95.2
werkzeug
uvicorn==0.22.0
psycopg2
sqlalchemy==2.0.21
python-dotenv==1.0.0
pytest==7.4.2
pyjwt==2.8.0
